% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_interactions.R
\name{annotate_interactions}
\alias{annotate_interactions}
\title{Annotates GenomicInteractions object from load_interactions}
\usage{
annotate_interactions(interactions, annotation, ...)
}
\arguments{
\item{interactions}{GenomicInteractions object from \code{\link{load_interactions}}}

\item{annotation}{full path to annotations file}

\item{...}{additional parameters for fread}
}
\value{
GenomicInteractions object annotated
}
\description{
This function annotate a GenomicInteractions object from load_interactions, based on a given annotations file
}
