% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_interactions.R
\name{.annotate_BOE}
\alias{.annotate_BOE}
\title{Annotates interactions into B_B or B_OE}
\usage{
.annotate_BOE(interactions)
}
\arguments{
\item{interactions}{HiCaptuRe object from \code{\link{load_interactions}}}
}
\value{
HiCaptuRe object with the column int updated
}
\description{
This internal function annotate a HiCaptuRe object based on the presence or not of annotation in a given restriction fragment
}
\keyword{internal}
