% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_windows.R
\name{expression_windows}
\alias{expression_windows}
\title{Extract counts for windows flanking a set of regions}
\usage{
expression_windows(regions, windows, dds, promoters)
}
\arguments{
\item{regions}{a GenomicRanges object}

\item{windows}{a sequence of windows starting from 0}

\item{dds}{a DESeqDataSet or DESeqTransform object}

\item{promoters}{a GenomicRanges with the promoter regions and the gene associated}
}
\value{
dataframe with 7 columns: gene_id, sample, count, cond, windows, total_genes, all_zero
1. 'gene_id' gene_id from the dds object
2. 'sample' name of the sample in the dds object
3. 'count' count value for the specific gene
4. 'cond' condition (name of sample without _1/_2)
5. 'windows' number of the window
6. 'total_genes' number of total genes in the specific windows
7. 'all_zero' number of genes that have 0 counts in all samples
}
\description{
This function extract the counts for the genes in windows flanking a set of given regions
}
