% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiCaptuRe.R
\docType{package}
\name{HiCaptuRe}
\alias{HiCaptuRe}
\title{HiCaptuRe Package}
\description{
In-house R package for Capture Hi-C data handling
}
\author{
Laureano Tomás-Daza \email{ltomas@carrerasresearch.org}
}
