% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_regions.R
\name{integrate_regions}
\alias{integrate_regions}
\title{Filters GenomicInteractions object by integrating a regions file}
\usage{
integrate_regions(
  interactions,
  regions,
  chr = NULL,
  start = NULL,
  end = NULL,
  invert = F
)
}
\arguments{
\item{interactions}{GenomicInteractions object from \code{\link{load_interactions}}}

\item{regions}{full path to regions file (bed format) or a GRanges object}

\item{chr}{column name of chromosome values}

\item{start}{column name of start positions}

\item{end}{column name of end positions}

\item{invert}{T/F if need those interactions that do NOT overlaps with any regions}
}
\value{
GenomicInteractions object filtered by regions with 'int' column containing type of interactions
}
\description{
This function filters a GenomicInteractions object from load_interactions by integrating regions from a external regions file, and classifies the interactions by type of interactions based on if the right, left or both end overlap with a region
}
