% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions2bedpe.R
\name{interactions2bedpe}
\alias{interactions2bedpe}
\title{Generate bedpe file}
\usage{
interactions2bedpe(interactions, file, over.write = F)
}
\arguments{
\item{interactions}{GenomicInteractions object from \code{\link{load_interactions}}}

\item{file}{full path to desired output file}

\item{over.write}{T/F to over write the output file}
}
\value{
tibble object with the bedpe table and save it in the desired output file
}
\description{
This function creates data.frame and files in ibed format
}
