% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions2cytoscape.R
\name{interactions2cytoscape}
\alias{interactions2cytoscape}
\title{Generate Cytoscape Network file}
\usage{
interactions2cytoscape(interactions, file, type = "complete")
}
\arguments{
\item{interactions}{GenomicInteractions object from \code{\link{load_interactions}}}

\item{file}{full path to desired output file}

\item{type}{Type of network you want to obtained (Complete, P_P or P_OE)}
}
\value{
File needed to upload to Cytoscape
}
\description{
This function creates the file needed to upload to Cytoscape
}
