% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions2washU.R
\name{interactions2washUold}
\alias{interactions2washUold}
\title{Generate file to upload to WashU Browser}
\usage{
interactions2washUold(interactions, file)
}
\arguments{
\item{interactions}{GenomicInteractions object from \code{\link{load_interactions}}}

\item{file}{full path to desired output file}
}
\value{
tibble object with the necessary table for the WashU Browser, and save it in the desired output file
}
\description{
This function creates data.frame and files to load as track in the WashU Browser
}
