% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions_summary.R
\name{interactions_summary}
\alias{interactions_summary}
\title{Generate a PDF report for the interaction object}
\usage{
interactions_summary(interactions, pdf_file, ...)
}
\arguments{
\item{interactions}{GenomicInteractions object from \code{\link{load_interactions}}}

\item{pdf_file}{file path to save the report}

\item{...}{additional parameters for split_by_distance}
}
\value{
PDF report
}
\description{
This function generates a PDF summary report of the GInteractions object
}
