% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactionsByBaits.R
\name{interactionsByBaits}
\alias{interactionsByBaits}
\title{Filters GenomicInteractions object by baits}
\usage{
interactionsByBaits(interactions, baits)
}
\arguments{
\item{interactions}{a GenomicInteractions object from \code{\link{load_interactions}}}

\item{baits}{character vector containing bait names}
}
\value{
GenomicInteractions object, subset of the original based on the give baits
}
\description{
his function filters a GenomicInteractions object from load_interactions by a set of given bait(s)
}
