% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactionsByGenes.R
\name{interactionsByGenes}
\alias{interactionsByGenes}
\title{Computes the intersect between a vector of genes and interactions}
\usage{
interactionsByGenes(interactions, genes, annotation, genes_annot)
}
\arguments{
\item{interactions}{a GenomicInteractions object from \code{\link{load_interactions}}}

\item{genes}{character vector containing genes}

\item{annotation}{full path to annotation file which MUST include header as follows: chr, start, end, and at least a colum with genes_annot}

\item{genes_annot}{column name of annation file with the annotation in the same format as the genes}
}
\value{
GenomicInteractions object, subset of the original based on the give genes
}
\description{
This function computes the intersect between a given vector of genes and a GenomicInteractions objects using an annotaion file
}
