% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactionsByRegions.R
\name{interactionsByRegions}
\alias{interactionsByRegions}
\title{Filters GenomicInteractions object by integrating regions}
\usage{
interactionsByRegions(
  interactions,
  regions,
  chr = NULL,
  start = NULL,
  end = NULL,
  invert = F
)
}
\arguments{
\item{interactions}{GenomicInteractions object from \code{\link{load_interactions}}}

\item{regions}{full path to regions file (bed format) or a GRanges object}

\item{chr}{column name of chromosome values}

\item{start}{column name of start positions}

\item{end}{column name of end positions}

\item{invert}{T/F if need those interactions that do NOT overlaps with any regions}
}
\value{
GenomicInteractions object filtered by regions, by default with additional columns regarding overlap on each node. If invert=T no additional columns.
}
\description{
This function filters a GenomicInteractions object from load_interactions by integrating regions and classifies the interactions by type of interactions based on if the right, left or both end overlap with a region
}
