% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_genome.R
\name{load_genome}
\alias{load_genome}
\title{Load genome}
\usage{
load_genome(genome)
}
\arguments{
\item{genome}{character with the name of the genome version}
}
\value{
BSgenome object
}
\description{
This function load the given genome if installed
}
