% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakmatrix2list.R
\name{peakmatrix2list}
\alias{peakmatrix2list}
\title{Creates a list of HiCaptuRe objects based on a peakmatrix}
\usage{
peakmatrix2list(peakmatrix, cutoff = 5)
}
\arguments{
\item{peakmatrix}{HiCaptuRe object of a peakmatrix}

\item{cutoff}{Chicago score cutoff to filter interactions}
}
\value{
list with HiCaptuRe objects
}
\description{
This function creates a list of HiCaptuRe objects separating all the samples included on a peakmatrix.
}
