% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distance.R
\name{plot_distance}
\alias{plot_distance}
\title{Generate the plot based on distance output}
\usage{
plot_distance(distances, type_of_value = "absolute")
}
\arguments{
\item{distances}{list of dataframes output from \code{\link{split_by_distance}}}

\item{type_of_value}{an element from: absolute, by_int_type, by_total}
}
\value{
list with the plots
}
\description{
This function plots the distances by different approaches for the output of split_by_distance
}
\details{
The type_of_value argument could be: Absolute: plots the Absolute number of interactions. By_int_type: plots the number of interactions divided by the total number of interactions of each type. By_total: plots the number of interactions divided by the total number of interactions
}
