% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_by_distance.R
\name{split_by_distance}
\alias{split_by_distance}
\title{Computes the number of interactions by distance}
\usage{
split_by_distance(interactions, cell_type = "cell_type")
}
\arguments{
\item{interactions}{GenomicInteractions object from \code{\link{load_interactions}}}

\item{cell_type}{character variable with the name of the cell type or sample}
}
\value{
list with 2 tables: short_int_dist_table and long_int_dist_table, with the number of short interactions and long interactions respectivetly
}
\description{
This function computes the number of interactions by distance for a given GenomicInteractions object
}
