% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getters}
\alias{getters}
\alias{getParameters}
\alias{getParameters,HiCaptuRe-method}
\alias{getByBaits}
\alias{getByBaits,HiCaptuRe-method}
\alias{getByRegions}
\alias{getByRegions,HiCaptuRe-method}
\title{Functions to access data held in a HiCaptuRe object.}
\usage{
getParameters(x)

\S4method{getParameters}{HiCaptuRe}(x)

getByBaits(x)

\S4method{getByBaits}{HiCaptuRe}(x)

getByRegions(x)

\S4method{getByRegions}{HiCaptuRe}(x)
}
\arguments{
\item{HiCaptuRe}{A HiCaptuRe object}
}
\value{
For 'getParameters', a named list of named vectors. For 'getByBaits', a list
of tibbles with bait-centric information. For 'getByRegions', a list of GRanges with
regions-centric information.
}
\description{
Use these functions to access data stored in each of the slots of a
HiCapture object.
}
