% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiCaptuRe-class.R
\docType{class}
\name{HiCaptuRe-class}
\alias{HiCaptuRe-class}
\title{HiCaptuRe Class}
\description{
A S4 class to represent interactions between genomic regions
}
\section{Slots}{

\describe{
\item{\code{parameters}}{List of parameters used to create the object and subsequence analysis}

\item{\code{ByBaits}}{List of Baits used by \link[HiCaptuRe]{interactionsByBaits}.}

\item{\code{ByRegions}}{List of Regions used by \link[HiCaptuRe]{interactionsByRegions}}
}}

\note{
This class contains a \link[GenomicInteractions]{GenomicInteractions} object inside therefore all methods available to it can be used. This type of object should be created through the function \link[HiCaptuRe]{load_interactions}
}
