% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiCaptuRe-package.R
\docType{package}
\name{HiCaptuRe-package}
\alias{HiCaptuRe}
\alias{HiCaptuRe-package}
\title{HiCaptuRe: In-house R package for Capture Hi-C data handling}
\description{
Package for dealing with liCHi-C data in Javierre Lab.
}
\keyword{internal}
