% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactionsByBaits.R
\name{interactionsByBaits}
\alias{interactionsByBaits}
\title{Filters HiCaptuRe object by baits}
\usage{
interactionsByBaits(interactions, baits, sep = ",")
}
\arguments{
\item{interactions}{a HiCaptuRe object}

\item{baits}{character vector containing bait names}

\item{sep}{character separating baits names when several baits in same fragment}
}
\value{
HiCaptuRe object, subset of the original based on the given baits. And an additional slot ByBatis with bait-centric statistics
}
\description{
This function filters a HiCaptuRe object from load_interactions by a set of given bait(s)
}
