% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactionsByRegions.R
\name{interactionsByRegions}
\alias{interactionsByRegions}
\title{Filters HiCaptuRe object by overlaping regions}
\usage{
interactionsByRegions(
  interactions,
  regions,
  chr = NULL,
  start = NULL,
  end = NULL,
  invert = F
)
}
\arguments{
\item{interactions}{HiCaptuRe object}

\item{regions}{full path to regions file (bed format) or a GRanges object}

\item{chr}{column name of chromosome values}

\item{start}{column name of start positions}

\item{end}{column name of end positions}

\item{invert}{T/F if need those interactions that do NOT overlaps with any regions}
}
\value{
HiCaptuRe object filtered by regions, by default with additional columns regarding overlap on each node. If invert=T no additional columns. And an additional slot ByRegions with region-centric statistics
}
\description{
This function filters a HiCaptuRe object from load_interactions by overlaping regions
}
