% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_interactions.R
\name{intersect_interactions}
\alias{intersect_interactions}
\title{Computes the intersect between a list of interactions objects}
\usage{
intersect_interactions(interactions_list, distance.boxplot = F, ...)
}
\arguments{
\item{interactions_list}{list of GenomicInteractions objects from \code{\link{load_interactions}}}

\item{distance.boxplot}{logical, by default FALSE, plot a boxplot of log10 distance of each intersection in the upset plot}

\item{...}{extra arguments for UpSetR}
}
\value{
list with 3 elements:
1. 'upset_plot'  an upset plot of the intersections
2. 'venn' a venn diagram of the intersections
3. 'intersections' a list of GenomicInteractions objects for all the intersections
}
\description{
This function computes all the possible intersects for a given list of GenomicInteractions objects
}
